Refactor and redesign
This change makes some substantial changes to the project:
* Create public interface for codegen in the pkg/codegen namespace
  while placing internal utilities under internal/
* Simplify cobra usage, simplify program config substantially
* A new test env var, TEST_SKIP_CLEANUP, can be set to instruct the
  crds_test.go tests to not perform temp dir cleanup after the test
  run, for the purposes of investigating bad output during test failure.
  Generated code is now placed in temp dirs with friendly, identifiable
  names for each test case.
* General refactoring: removal of dead code, reorganizing functions into
  more appropriately named files or packages.
* Update to latest Pulumi SDK as well as update all other dependencies.
* Update to Go 1.18

+++ tests/crds_test.go
@@ -15,14 +15,16 @@
 package tests
 
 import (
+	"fmt"
 	"io/fs"
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 var languages = []string{"dotnet", "go", "nodejs", "python"}
@@ -31,19 +33,25 @@ const gkeManagedCertsUrl = "https://raw.githubusercontent.com/GoogleCloudPlatfor
 
 // execCrd2Pulumi runs the crd2pulumi binary in a temporary directory
 func execCrd2Pulumi(t *testing.T, lang, path string) {
-	tmpdir, err := ioutil.TempDir("", "")
+	dir := filepath.Join(os.TempDir(), fmt.Sprintf("crd2pulumi-tests-%s-%s", time.Now().Local().Format(time.Kitchen), lang))
+	require.NoError(t, os.MkdirAll(dir, 0755))
+	tmpdir, err := os.MkdirTemp(dir, t.Name()+"-*")
 	assert.Nil(t, err, "expected to create a temp dir for the CRD output")
-	defer os.RemoveAll(tmpdir)
-	langFlag := "--" + lang + "Path"
+	if os.Getenv("TEST_SKIP_CLEANUP") != "" {
+		t.Logf("Test output tmp dir: %q\n", tmpdir)
+	} else {
+		defer os.RemoveAll(tmpdir)
+	}
 	binaryPath, err := filepath.Abs("../bin/crd2pulumi")
 	if err != nil {
 		panic(err)
 	}
-	t.Logf("%s %s=%s %s: running", binaryPath, langFlag, tmpdir, path)
-	crdCmd := exec.Command(binaryPath, langFlag, tmpdir, "--force", path)
+	args := []string{"--lang", lang, "--outputDir", tmpdir, "--force", path}
+	t.Logf("running %q %q", binaryPath, args)
+	crdCmd := exec.Command(binaryPath, args...)
 	crdOut, err := crdCmd.CombinedOutput()
-	t.Logf("%s %s=%s %s: output=\n%s", binaryPath, langFlag, tmpdir, path, crdOut)
-	assert.Nil(t, err, "expected crd2pulumi for '%s=%s %s' to succeed", langFlag, tmpdir, path)
+	t.Logf("output=\n%s\n", crdOut)
+	assert.Nil(t, err, "expected no error running crd2pulumi")
 }
 
 // TestCRDsFromFile enumerates all CRD YAML files, and generates them in each language.
